requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Authlib==1.6.0
babel==2.17.0
bashlex==0.18
beautifulsoup4==4.13.4
bidict==0.23.1
binaryornot==0.4.4
bitarray==3.4.2
bleach==6.2.0
boto3==1.38.33
botocore==1.38.33
browsergym-core==0.13.3
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ckzg==2.1.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
clr_loader==0.2.7.post0
cobble==0.1.4
comm==0.2.2
contourpy==1.3.2
crashtest==0.4.1
cryptography==45.0.3
cycler==0.12.1
cytoolz==1.0.1
daytona_api_client==0.20.1
daytona_api_client_async==0.20.1
daytona_sdk==0.20.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirhash==0.5.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
dulwich==0.22.8
e2b==1.5.1
environs==9.5.0
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
fastmcp==2.6.1
filelock==3.18.0
findpython==0.6.3
flake8==7.2.0
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.96.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.18.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grep-ast==0.9.0
groq==0.27.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.62.3
grpclib==0.4.7
gymnasium==1.1.1
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hf-xet==1.1.2
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.1.0
installer==0.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
json_repair==0.46.2
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-kernel-gateway==3.0.1
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.8
libcst==1.5.0
libtmux==0.39.0
litellm==1.72.2
lxml==5.4.0
mammoth==1.9.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.9.2
mdurl==0.1.2
memory-profiler==0.61.0
minio==7.2.15
mistune==3.1.3
modal==0.77.0
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.2
openai==1.85.0
openapi-pydantic==0.5.1
openhands-aci==0.3.0
openpyxl==3.1.5
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pbs-installer==2025.5.17
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
playwright==1.52.0
poetry==2.1.3
poetry-core==2.1.3
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pyflakes==3.3.2
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-engineio==4.12.1
python-frontmatter==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socketio==5.13.0
python-telegram-bot==22.1
pythonnet==3.0.5
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
pyzmq==26.4.0
qtconsole==5.6.1
QtPy==2.4.3
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rlp==4.1.0
rpds-py==0.25.1
rsa==4.9.1
runloop_api_client==0.39.0
s3transfer==0.13.0
scantree==0.0.4
schedule==1.2.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
sigtools==4.0.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
stripe==12.2.0
synchronicity==0.9.12
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.23.2
tree-sitter-language-pack==0.7.3
tree-sitter-yaml==0.7.1
trove-classifiers==2025.5.9.12
typer==0.16.0
types-certifi==2021.10.8.3
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.31.2
watchfiles==1.0.5
wcwidth==0.2.13
web3==7.12.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whatthepatch==1.0.7
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.3
yarl==1.20.0
youtube-transcript-api==1.0.3
zipp==3.22.0
zope.interface==7.2
zstandard==0.23.0
